AMGCLWrap

Julia wrapper around AMGCL via the C wrapper AMGCL_C.

If you find this package useful, please give credits to the author of AMGCL.

AMGCLWrap.AMGSolver — Type

AMGSolver(sparsematrix; blocksize=1, param=nothing)

Create Algebraic multigrid preconditioned Krylov subspace solver with ldiv! and \ methods solving the matrix system.

Input:

sparsematrix: SparseArrays.AbstractSparseMatrixCSC or SparseMatricesCSR.SparseMatrixCSR.

blocksize: If blocksize >1, group unknowns into blocks of given size and cast the matrix internally to a sparse matrix of blocksize x blocksize static matrices. Block sizes 1...8 are instantiated.

params: Any object (e.g. Tuple, Dict or JSON string) which can be turned into a JSON string by JSON3.write. If params is an emtpy string or nothing a default value is used.

AMGCLWrap.RLXSolver — Type

RLXSolver(sparsematrix; blocksize=1, param=nothing)

Create single level relaxation preconditioned Krylov subspace solver with ldiv! and \ methods solving the matrix system.

Input:

sparsematrix: SparseArrays.AbstractSparseMatrixCSC or SparseMatricesCSR.SparseMatrixCSR.

blocksize: If blocksize >1, group unknowns into blocks of given size and cast the matrix internally to a sparse matrix of blocksize x blocksize static matrices. Block sizes 1...8 are instantiated.

params: Any object (e.g. Tuple, Dict or JSON string) which can be turned into a JSON string by JSON3.write. If params is an emtpy string or nothing a default value is used.

AMGCLWrap.AMGPrecon — Type

AMGPrecon(sparsematrix; blocksize=1, param=nothing)

Create algebraic multigrid preconditioner with ldiv! and \ methods solving the preconditioning system.

Input:

sparsematrix: SparseArrays.AbstractSparseMatrixCSC or SparseMatricesCSR.SparseMatrixCSR.

blocksize: If blocksize >1, group unknowns into blocks of given size and cast the matrix internally to a sparse matrix of blocksize x blocksize static matrices. Block sizes 1...8 are instantiated.

params: Any object (e.g. Tuple, Dict or JSON string) which can be turned into a JSON string by JSON3.write. If params is an emtpy string or nothing a default value is used.

AMGCLWrap.RLXPrecon — Type

RLXPrecon(sparsematrix; blocksize=1, param=nothing)

Create single level relaxation preconditioner with ldiv! and \ methods solving the preconditioning system.

Input:

sparsematrix: SparseArrays.AbstractSparseMatrixCSC or SparseMatricesCSR.SparseMatrixCSR.

blocksize: If blocksize >1, group unknowns into blocks of given size and cast the matrix internally to a sparse matrix of blocksize x blocksize static matrices. Block sizes 1...8 are instantiated.

params: Any object (e.g. Tuple, Dict or JSON string) which can be turned into a JSON string by JSON3.write. If params is an emtpy string or nothing a default value is used.