Parameters

AMGCLWrap.AbstractAMGCLParams — Type

abstract type AbstractAMGCLParams

Abstract parameter type.

Iterative solver strategies

See the Iterative Solvers section of the AMGCL documentation.

Instead of one of the solvers below, a named tuple like (type= "cg", tol=1.0e-10) can be passed. This also allows to use methods not accessible via one of the structures defined below.

AMGCLWrap.AbstractSolver — Type

abstract type AbstractSolver <: AMGCLWrap.AbstractAMGCLParams

Abstract solver parameter type.

AMGCLWrap.BICGStabSolver — Type

struct BICGStabSolver <: AMGCLWrap.AbstractSolver

BICGStab solver

type::String

pside::String

tol::Float64

abstol::Float64

maxiter::Int64

verbose::Bool

AMGCLWrap.CGSolver — Type

struct CGSolver <: AMGCLWrap.AbstractSolver

type::String

tol::Float64

abstol::Float64

maxiter::Int64

verbose::Bool

AMGCLWrap.GMRESSolver — Type

struct GMRESSolver <: AMGCLWrap.AbstractSolver

type::String

M::Int64

pside::String

tol::Float64

abstol::Float64

maxiter::Int64

verbose::Bool

Relaxation strategies

See the Relaxation section of the AMGCL documentation.

Instead of one of the strategies below, a named tuple like (type= "damped_jacobi", damping=0.72) can be passed. This also allows to use methods not accessible via one of the structures defined below.

AMGCLWrap.AbstractRelaxation — Type

abstract type AbstractRelaxation <: AMGCLWrap.AbstractAMGCLParams

Abstract relaxation parameter type.

AMGCLWrap.ILU0Relaxation — Type

struct ILU0Relaxation <: AMGCLWrap.AbstractRelaxation

type::String

AMGCLWrap.SPAI0Relaxation — Type

struct SPAI0Relaxation <: AMGCLWrap.AbstractRelaxation

type::String

Coarsening strategies

See the Coarsening Strategies section of the AMGCL documentation.

Instead of one of the strategies below, a named tuple like (type= "smoothed_aggregation", relax=1.0) can be passed. This also allows to use methods not accessible via one of the structures defined below.

AMGCLWrap.AbstractCoarsening — Type

abstract type AbstractCoarsening <: AMGCLWrap.AbstractAMGCLParams

Abstract coarsening parameter type.

AMGCLWrap.RugeStubenCoarsening — Type

struct RugeStubenCoarsening <: AMGCLWrap.AbstractCoarsening

type::String

eps_strong::Float64

do_trunc::Bool

eps_trunc::Float64

AMGCLWrap.SmoothedAggregationCoarsening — Type

struct SmoothedAggregationCoarsening <: AMGCLWrap.AbstractCoarsening

type::String

relax::Float64