AMGCLWrap

Julia wrapper around a subset of AMGCL via the C wrapper AMGCL_C.

If you find this package useful, please give credits to the author of AMGCL.

Solvers

AMGCLWrap.AMGSolver — Type

AMGSolver(sparsematrix::AbstractSparseMatrix; 
             blocksize=1, 
             param=nothing)

Create Algebraic multigrid preconditioned Krylov subspace solver with ldiv! and \ methods solving the matrix system.

Parameters:

sparsematrix: SparseArrays.AbstractSparseMatrixCSC or SparseMatricesCSR.SparseMatrixCSR.

blocksize: If blocksize >1, group unknowns into blocks of given size and cast the matrix internally to a sparse matrix of blocksize x blocksize static matrices. Block sizes 1...8 are instantiated.

param: Any object (e.g. Tuple, Dict or JSON string) which can be turned into a JSON string by JSON3.write. If params is an emtpy string or nothing a default value is used.

AMGSolver(sparsematrix::AbstractSparseMatrix;
          blocksize::Int=1,
          param=nothing,
          verbose::Bool=false,
          coarsening::Union{AbstractCoarsening, NamedTuple}=SmoothedAggregationCoarsening(),
          relax::Union{AbstractRelaxation, NamedTuple}=SPAI0Relaxation(),
          solver::Union{AbstractSolver, NamedTuple}=BICGStabSolver(;verbose))

Create Algebraic multigrid preconditioned Krylov subspace solver with ldiv! and \ methods solving the matrix system.

Parameters:

sparsematrix: SparseArrays.AbstractSparseMatrixCSC or SparseMatricesCSR.SparseMatrixCSR.

blocksize: If blocksize >1, group unknowns into blocks of given size and cast the matrix internally to a sparse matrix of blocksize x blocksize static matrices. Block sizes 1...8 are instantiated.

param: Ignored if nothing (default). Otherwise, any object (e.g. Tuple, Dict or JSON string) which can be turned into a JSON string by JSON3.write.

verbose: if true, print generated JSON string passed to amgcl.

coarsening: A coarsening strategy

relax: A relaxation method

solver: An iterative solver

AMGCLWrap.RLXSolver — Type

RLXSolver(sparsematrix::AbstractSparseMatrix; 
             blocksize=1, 
             param=nothing)

Create single level relaxation preconditioned Krylov subspace solver with ldiv! and \ methods solving the matrix system.

Parameters:

sparsematrix: SparseArrays.AbstractSparseMatrixCSC or SparseMatricesCSR.SparseMatrixCSR.

blocksize: If blocksize >1, group unknowns into blocks of given size and cast the matrix internally to a sparse matrix of blocksize x blocksize static matrices. Block sizes 1...8 are instantiated.

param: Any object (e.g. Tuple, Dict or JSON string) which can be turned into a JSON string by JSON3.write. If params is an emtpy string or nothing a default value is used.

RLXSolver(sparsematrix::AbstractSparseMatrix;
          blocksize::Int=1,
          param=nothing,
          verbose::Bool=false,
          precond::Union{AbstractRelaxation, NamedTuple}=SPAI0Relaxation(),
          solver::Union{AbstractSolver, NamedTuple}=BICGStabSolver(;verbose))

Create single level relaxation preconditioned Krylov subspace solver with ldiv! and \ methods solving the matrix system.

Parameters:

sparsematrix: SparseArrays.AbstractSparseMatrixCSC or SparseMatricesCSR.SparseMatrixCSR.

blocksize: If blocksize >1, group unknowns into blocks of given size and cast the matrix internally to a sparse matrix of blocksize x blocksize static matrices. Block sizes 1...8 are instantiated.

param: Ignored if nothing (default). Otherwise, any object (e.g. Tuple, Dict or JSON string) which can be turned into a JSON string by JSON3.write.

verbose: if true, print generated JSON string passed to amgcl.

precond: A preconditioned method

solver: An iterative solver

Preconditioners

AMGCLWrap.AMGPrecon — Type

AMGPrecon(sparsematrix::AbstractSparseMatrix; 
             blocksize=1, 
             param=nothing)

Create algebraic multigrid preconditioner with ldiv! and \ methods solving the preconditioning system.

Parameters:

sparsematrix: SparseArrays.AbstractSparseMatrixCSC or SparseMatricesCSR.SparseMatrixCSR.

blocksize: If blocksize >1, group unknowns into blocks of given size and cast the matrix internally to a sparse matrix of blocksize x blocksize static matrices. Block sizes 1...8 are instantiated.

param: Any object (e.g. Tuple, Dict or JSON string) which can be turned into a JSON string by JSON3.write. If params is an emtpy string or nothing a default value is used.

AMGPrecon(sparsematrix::AbstractSparseMatrix; blocksize::Int=1, param=nothing, verbose::Bool=false, coarsening::Union{AbstractCoarsening, NamedTuple}=SmoothedAggregationCoarsening(), relax::Union{AbstractRelaxation, NamedTuple}=SPAI0Relaxation()) Create algebraic multigrid preconditioner with ldiv! and \ methods solving the preconditioning system.

Parameters:

sparsematrix: SparseArrays.AbstractSparseMatrixCSC or SparseMatricesCSR.SparseMatrixCSR.

blocksize: If blocksize >1, group unknowns into blocks of given size and cast the matrix internally to a sparse matrix of blocksize x blocksize static matrices. Block sizes 1...8 are instantiated.

param: Ignored if nothing (default). Otherwise, any object (e.g. Tuple, Dict or JSON string) which can be turned into a JSON string by JSON3.write.

verbose: if true, print generated JSON string passed to amgcl.

coarsening: A coarsening strategy

relax: A relaxation method

AMGCLWrap.RLXPrecon — Type

RLXPrecon(sparsematrix::AbstractSparseMatrix; 
             blocksize=1, 
             param=nothing)

Create single level relaxation preconditioner with ldiv! and \ methods solving the preconditioning system.

Parameters:

sparsematrix: SparseArrays.AbstractSparseMatrixCSC or SparseMatricesCSR.SparseMatrixCSR.

blocksize: If blocksize >1, group unknowns into blocks of given size and cast the matrix internally to a sparse matrix of blocksize x blocksize static matrices. Block sizes 1...8 are instantiated.

param: Any object (e.g. Tuple, Dict or JSON string) which can be turned into a JSON string by JSON3.write. If params is an emtpy string or nothing a default value is used.

RLXPrecon(sparsematrix::AbstractSparseMatrix;
          blocksize::Int=1,
          param=nothing,
          verbose::Bool=false,
          precond::Union{AbstractRelaxation, NamedTuple}=SPAI0Relaxation())

Create single level relaxation preconditioner with ldiv! and \ methods solving the preconditioning system.

Parameters:

sparsematrix: SparseArrays.AbstractSparseMatrixCSC or SparseMatricesCSR.SparseMatrixCSR.

blocksize: If blocksize >1, group unknowns into blocks of given size and cast the matrix internally to a sparse matrix of blocksize x blocksize static matrices. Block sizes 1...8 are instantiated.

param: Ignored if nothing (default). Otherwise, any object (e.g. Tuple, Dict or JSON string) which can be turned into a JSON string by JSON3.write.

verbose: if true, print generated JSON string passed to amgcl.

precond: A preconditioned method

General tools

AMGCLWrap.error_state — Function

error_state(operator)::Int

Return error state of operator. Ok if 0.

AMGCLWrap.blocksize_instantiated — Function

blocksize_instantiated(blocksize)::Bool

Check if given blocksize has been instantiated.

Parameters

AMGCLWrap.AbstractAMGCLParams — Type

abstract type AbstractAMGCLParams

Abstract parameter type.

Solver parameters

See the Iterative Solvers section of the AMGCL documentation.

Instead of one of the solvers below, a named tuple like (type= "cg", tol=1.0e-10) can be passed This also allows to use methods not accessible via one of the structures defined below.

AMGCLWrap.AbstractSolver — Type

abstract type AbstractSolver <: AMGCLWrap.AbstractAMGCLParams

Abstract solver parameter type.

AMGCLWrap.BICGStabSolver — Type

struct BICGStabSolver <: AMGCLWrap.AbstractSolver

BICGStab solver

type::String

pside::String

tol::Float64

abstol::Float64

maxiter::Int64

verbose::Bool

AMGCLWrap.CGSolver — Type

struct CGSolver <: AMGCLWrap.AbstractSolver

type::String

tol::Float64

abstol::Float64

maxiter::Int64

verbose::Bool

AMGCLWrap.GMRESSolver — Type

struct GMRESSolver <: AMGCLWrap.AbstractSolver

type::String

M::Int64

pside::String

tol::Float64

abstol::Float64

maxiter::Int64

verbose::Bool

Relaxation/Preconditioner parameters

See the Relaxation section of the AMGCL documentation.

Instead of one of the strategies below, a named tuple like (type= "damped_jacobi", damping=0.72) can be passed This also allows to use methods not accessible via one of the structures defined below.

AMGCLWrap.AbstractRelaxation — Type

abstract type AbstractRelaxation <: AMGCLWrap.AbstractAMGCLParams

Abstract relaxation parameter type.

AMGCLWrap.ILU0Relaxation — Type

struct ILU0Relaxation <: AMGCLWrap.AbstractRelaxation

type::String

AMGCLWrap.SPAI0Relaxation — Type

struct SPAI0Relaxation <: AMGCLWrap.AbstractRelaxation

type::String

Coarsening strategies

See the Coarsening Strategies section of the AMGCL documentation.

Instead of one of the strategies below, a named tuple like (type= "smoothed_aggregation", relax=1.0) can be passed This also allows to use methods not accessible via one of the structures defined below.

AMGCLWrap.AbstractCoarsening — Type

abstract type AbstractCoarsening <: AMGCLWrap.AbstractAMGCLParams

Abstract coarsening parameter type.

AMGCLWrap.RugeStubenCoarsening — Type

struct RugeStubenCoarsening <: AMGCLWrap.AbstractCoarsening

type::String

eps_strong::Float64

do_trunc::Bool

eps_trunc::Float64

AMGCLWrap.SmoothedAggregationCoarsening — Type

struct SmoothedAggregationCoarsening <: AMGCLWrap.AbstractCoarsening

type::String

relax::Float64